+++ drivers-and-interrupts/utils.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <linux/types.h>
 
 typedef struct s_list
 {
@@ -13,4 +13,4 @@ void	ft_lstclear(t_list **alst, void (*del)(const void *));
 void	ft_lstadd_back(t_list **alst, t_list *new);
 void	ft_lstadd_front(t_list **alst, t_list *start);
 char	*ft_strdup(const char *src);
-unsigned long int	ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
